STG local markers Run STG to find a set of genes that separate a given DA region from a local subset of cells.

STG local markers Run STG to find a set of genes that separate a given DA region from a local subset of cells.

STGlocalMarkers(X, da.regions, da.region.to.run, cell.label.info,
  cell.label.to.run, ...)

Arguments

X
matrix, normalized expression matrix of all cells in the dataset, genes are in rows, rownames must be gene names

da.regions
output from the function getDAregion()

da.region.to.run
numeric, which (single) DA region to find local markers for

cell.label.info
vector, cell labeling information to select the local subset of cells to compare with input DA region

cell.label.to.run
cell label(s) to select from cell.label.info that represent the local neiborhood for the input DA region

lambda
numeric, regularization parameter that weights the number of selected genes, a larger lambda leads to fewer genes, default 1.5

n.runs
integer, number of runs to run the model, default 5

python.use
character string, the Python to use, default "/usr/bin/python"

return.model
a logical value to indicate whether to return the actual model of STG

GPU
which GPU to use, default '', using CPU

Value

a list of results:

markers

a list of data.frame with markers for each DA region

accuracy

a numeric vector showing mean accuracy for each DA region

model

a list of model for each DA region, each model contains:
features

features used to train the model

selected.features

the selected features of the final run

pred

the linear prediction value for each cell from the model