DA-seq Step 3: get DA regions

Cluster the DA cells retained from Step 1 and Step 2 of DA-seq to obtain spatially coherent DA regions.

getDAregion(X, da.cells, cell.labels, labels.1, labels.2,
  prune.SNN = 1/15, resolution = 0.05, group.singletons = F,
  min.cell = 15, do.plot = T, plot.embedding = NULL, size = 0.5,
  ...)

Arguments

X
size N-by-p matrix, input merged dataset of interest after dimension reduction

da.cells
output from getDAcells() or updateDAcells()

cell.labels
size N vector, labels for each input cell

labels.1
vector, label name(s) that represent condition 1

labels.2
vector, label name(s) that represent condition 2

prune.SNN
parameter for Seurat function FindNeighbors(), default 1/15

resolution
parameter for Seurat function FindClusters(), default 0.05

group.singletons
parameter for Seurat function FindClusters(), default True

min.cell
integer, number of cells below which a DA region will be removed as outliers, default 15

do.plot
a logical value to indicate whether to return ggplot objects showing the results, default True

plot.embedding
size N-by-2 matrix, 2D embedding for the cells

size
cell size to use in the plot, default 0.5

...
other parameters to pass to Seurat FindClusters()

Value

a list of results

da.region.label

DA region label for each cell from the whole dataset, '0' represents non-DA cells.

DA.stat

a table showing DA score and p-value for each DA region

da.region.plot

ggplot object showing DA regions on plot.embedding